+++ Assets/Game/Body/Mortal/DeadHealth.cs
@@ -2,18 +2,18 @@ namespace Shooter.Game.Body
 {
     public sealed class DeadHealth : Health
     {
-        public override int Hp => 0;
+        public override double Hp => 0;
 
-        public override int MaxHp => 100;
+        public override double MaxHp => 100;
 
         public override bool Alive => false;
 
-        protected override void DamageRaw(int amount)
+        protected override void DamageRaw(double amount)
         {
 
         }
 
-        public override void Heal(int amount)
+        public override void Heal(double amount)
         {
         }
