+++ Assets/Game/Body/Mortal/ImmortalHealth.cs
@@ -2,18 +2,18 @@ namespace Shooter.Game.Body
 {
     public sealed class ImmortalHealth : Health
     {
-        public override int Hp => 100;
+        public override double Hp => 100;
 
-        public override int MaxHp => 100;
+        public override double MaxHp => 100;
 
         public override bool Alive => true;
 
-        protected override void DamageRaw(int amount)
+        protected override void DamageRaw(double amount)
         {
 
         }
 
-        public override void Heal(int amount)
+        public override void Heal(double amount)
         {
         }
